penghapusan taxfunction
penghapusan class taxfunction dan memindahkan semua logika perhitungan tax sepenuhnya di class taxcalculator
