+++ src/taxcalculator/employee/Child.java
@@ -3,17 +3,17 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package taxcalculator;
+package taxcalculator.employee;
 
 /**
  *
  * @author Farrel Haykal
  */
-public class Spouse {
+public class Child {
     private String name;
     private String idNumber;
 
-    public Spouse(String name, String idNumber) {
+    public Child(String name, String idNumber) {
         this.name = name;
         this.idNumber = idNumber;
     }
