+++ src/taxcalculator/employee/Spouse.java
@@ -1,0 +1,20 @@
+/*
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
+ */
+package taxcalculator.employee;
+
+/**
+ *
+ * @author Farrel Haykal
+ */
+public class Spouse {
+    private String name;
+    private String idNumber;
+
+    public Spouse(String name, String idNumber) {
+        this.name = name;
+        this.idNumber = idNumber;
+    }
+}
